[245] Improve the reusability of the code
Bug: https://github.com/svalyn/svalyn-studio/issues/245
